refactor: modularize front-end
